// src/app/features/user/models/user.update.ts
export interface UserUpdate {
    userId: string;
    avatarDocumentId: string;
    email: string;
    phoneNumber: string | null;
    name: string|null;
    ngaySinh: string;
    gioiTinh: number;
    diaChi: string | null;
    maTinh: string | null;
    maHuyen: string | null;
    maXa: string | null;
}
  